Refactor forms.
