+++ user_map/forms/__init__.py
@@ -1,2 +1,1 @@
-from login import *
 from user import *
