+++ user_map/forms/user.py
@@ -1,16 +1,13 @@
 # coding=utf-8
 """Django forms for User related routines."""
 from django.contrib.gis import forms
-from django.contrib.auth.forms import (
-    PasswordChangeForm,
-    PasswordResetForm,
-    SetPasswordForm)
+from django.contrib.auth.forms import PasswordResetForm
 from leaflet.forms.widgets import LeafletWidget
 
 from user_map.models import User, Role
 
 
-class UserForm(forms.ModelForm):
+class RegistrationForm(forms.ModelForm):
     """Form for user model."""
     name = forms.CharField(
         required=True,
@@ -60,7 +57,7 @@ class UserForm(forms.ModelForm):
 
     def clean(self):
         """Verifies that the values entered into the password fields match."""
-        cleaned_data = super(UserForm, self).clean()
+        cleaned_data = super(RegistrationForm, self).clean()
         if 'password' in cleaned_data and 'password2' in cleaned_data:
             if cleaned_data['password'] != cleaned_data['password2']:
                 raise forms.ValidationError(
@@ -73,11 +70,33 @@ class UserForm(forms.ModelForm):
         :param commit: Whether committed to db or not.
         :type commit: bool
         """
-        user = super(UserForm, self).save(commit=False)
+        user = super(RegistrationForm, self).save(commit=False)
         user.set_password(self.cleaned_data['password'])
         if commit:
             user.save()
         return user
+
+
+class LoginForm(forms.Form):
+    """Form for user to log in."""
+    class Meta:
+        """Meta of the form."""
+        fields = ['email', 'password']
+
+    email = forms.EmailField(
+        widget=forms.EmailInput(
+            attrs={
+                'class': 'form-control',
+                'placeholder': 'john@doe.com',
+            })
+    )
+    password = forms.CharField(
+        widget=forms.PasswordInput(
+            attrs={
+                'class': 'form-control',
+                'placeholder': 'Your s3cr3T password'
+            })
+    )
 
 
 class BasicInformationForm(forms.ModelForm):
